backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["zhiweit <theanzhiwei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.3"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.5"
sqlalchemy = "^2.0.20"
psycopg2-binary = "^2.9.7"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
